backend/requirements.txt
Flask
FastAPI
SQLAlchemy
psycopg2-binary
uvicorn
python-multipart
pydantic
alembic
passlib[bcrypt]
python-dotenv
pytest
httpx